+++ azure-storage-blob/ChangeLog.md
@@ -2,8 +2,9 @@
 
 > See [BreakingChanges](BreakingChanges.md) for a detailed list of API breaks.
 
-## Version XX.XX.XX:
+## Version 1.3.0:
 
+- Support for 2018-03-28 REST version. Please see our REST API documentation and blog for information about the related added features.
 - Added support for setting static website service properties.
 - Added support for getting account information, such as SKU name and account kind.
 - Added support for put block from URL(synchronously).